search screen and home screens seperated
