--- mobile/app/src/main/java/lk/eclk/locationservice/ui/home/HomeMessageEvents.kt
@@ -1,7 +1,0 @@
-package lk.eclk.locationservice.ui.home
-
-import lk.eclk.locationservice.internal.eventexecutor.MessageEvents
-
-interface HomeMessageEvents {
-    fun initRecyclerView(items: List<LocationListItem>)
-}
