+++ mobile/app/src/main/java/lk/eclk/locationservice/ui/search/SearchMessageEvents.kt
@@ -1,0 +1,5 @@
+package lk.eclk.locationservice.ui.search
+
+interface SearchMessageEvents {
+    fun initRecyclerView(items: List<LocationListItem>)
+}
