+++ mobile/app/src/main/java/lk/eclk/locationservice/ui/search/SearchViewModelFactory.kt
@@ -1,14 +1,14 @@
-package lk.eclk.locationservice.ui.home
+package lk.eclk.locationservice.ui.search
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import lk.eclk.locationservice.data.repository.Repository
 
-class HomeViewModelFactory(private val repository: Repository) :
+class SearchViewModelFactory(private val repository: Repository) :
     ViewModelProvider.NewInstanceFactory() {
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        return HomeViewModel(repository) as T
+        return SearchViewModel(repository) as T
     }
 }
