+++ mobile/app/src/main/res/navigation/main_navigation.xml
@@ -22,7 +22,7 @@
             app:popUpToInclusive="true" />
         <action
             android:id="@+id/action_splashScreenFragment_to_homeFragment"
-            app:destination="@id/homeFragment"
+            app:destination="@id/searchFragment"
             app:enterAnim="@anim/slide_in_down"
             app:exitAnim="@anim/fade_out"
             app:popEnterAnim="@anim/fade_in"
@@ -37,7 +37,7 @@
         tools:layout="@layout/sign_in_fragment">
         <action
             android:id="@+id/action_signInFragment_to_homeFragment"
-            app:destination="@id/homeFragment"
+            app:destination="@id/searchFragment"
             app:exitAnim="@anim/fade_out"
             app:popEnterAnim="@anim/fade_in"
             app:popExitAnim="@anim/slide_out_down"
@@ -45,8 +45,8 @@
             app:popUpToInclusive="true" />
     </fragment>
     <fragment
-        android:id="@+id/homeFragment"
-        android:name="lk.eclk.locationservice.ui.home.HomeFragment"
+        android:id="@+id/searchFragment"
+        android:name="lk.eclk.locationservice.ui.search.SearchFragment"
         android:label="home_fragment"
         tools:layout="@layout/home_fragment" />
